Organize imports so that cargo fmt can lint them
Order imports into groups (IntelliJ and RustRover use this convention):

  - `std::`
  - external
  - workspace
  - `crate::`
  - relative (`super::`/`self::`)

Do not curly-brace group imports from separate modules.

Prefer crate imports to relative imports, per the Rust Book.
The only exception is `use super::*` from test modules.

Remove unused imports.

Avoid re-exports (`pub use`).
