+++ attestation-transformer/src/schemas/status.rs
@@ -1,8 +1,10 @@
-use crate::did::Schema;
-use crate::{did::Did, error::AttTrError, term::Term, utils::address_from_ecdsa_key};
 use serde_derive::{Deserialize, Serialize};
 
-use super::{Domain, IntoTerm, Proof, Validation};
+use crate::did::{Did, Schema};
+use crate::error::AttTrError;
+use crate::schemas::{Domain, IntoTerm, Proof, Validation};
+use crate::term::Term;
+use crate::utils::address_from_ecdsa_key;
 
 #[derive(Deserialize, Serialize, Clone)]
 pub enum CurrentStatus {
@@ -101,12 +103,15 @@ impl IntoTerm for StatusSchema {
 
 #[cfg(test)]
 mod test {
-	use crate::schemas::status::{CredentialSubject, StatusSchema};
+	use secp256k1::rand::thread_rng;
+	use secp256k1::{generate_keypair, Message, Secp256k1};
+	use sha3::{Digest, Keccak256};
+
+	use crate::did::Did;
 	use crate::schemas::{Proof, Validation};
 	use crate::utils::address_from_ecdsa_key;
-	use crate::{did::Did, schemas::status::CurrentStatus};
-	use secp256k1::{generate_keypair, rand::thread_rng, Message, Secp256k1};
-	use sha3::{Digest, Keccak256};
+
+	use super::*;
 
 	#[test]
 	fn should_validate_endorse_credential() {
