+++ attestation-transformer/src/term.rs
@@ -1,5 +1,6 @@
-use crate::error::AttTrError;
 use proto_buf::transformer::{Form, TermObject};
+
+use crate::error::AttTrError;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 enum TermForm {
@@ -178,7 +179,7 @@ impl From<Term> for TermObject {
 
 #[cfg(test)]
 mod test {
-	use super::{Term, TermForm};
+	use super::*;
 
 	#[test]
 	fn should_convert_term_to_bytes_and_back() {
