+++ indexer/src/storage/lm_db/mod.rs
@@ -1,8 +1,10 @@
-use crate::storage::lm_db::types::LMDBClientConfig;
-use crate::storage::types::BaseKVStorage;
+use std::fs;
+
 use heed::types::Str;
 use heed::{Database, EnvOpenOptions};
-use std::fs;
+
+use crate::storage::lm_db::types::LMDBClientConfig;
+use crate::storage::types::BaseKVStorage;
 
 pub mod types;
