+++ indexer/src/tasks/csv_poc/task.rs
@@ -1,13 +1,14 @@
+use std::time::Duration;
+
 use digest::Digest;
 use hex;
 use serde::{Deserialize, Serialize};
 use serde_json;
 use sha3::Sha3_256;
-use std::time::Duration;
 use tracing::{debug, info};
 
-pub use crate::clients::csv::client::CSVClient;
-pub use crate::tasks::types::{BaseTask, BaseTaskState, TaskResponse};
+use crate::clients::csv::client::CSVClient;
+use crate::tasks::types::{BaseTask, BaseTaskState, TaskResponse};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct CSVPOCTaskState {
