+++ indexer/src/tasks/types.rs
@@ -1,5 +1,6 @@
-use serde::{Deserialize, Serialize};
 use std::time::Duration;
+
+use serde::{Deserialize, Serialize};
 
 // todo better layer separation
 #[tonic::async_trait]
