+++ job-manager/src/main.rs
@@ -1,14 +1,16 @@
-use proto_buf::combiner::linear_combiner_client::LinearCombinerClient;
-use proto_buf::combiner::{LtBatch, LtHistoryBatch};
-use proto_buf::transformer::transformer_client::TransformerClient;
-use proto_buf::transformer::{EventBatch, TermBatch};
 use std::error::Error;
 use std::time::Duration;
+
 use tokio::time::interval;
 use tokio_stream::wrappers::IntervalStream;
 use tokio_stream::StreamExt;
 use tonic::transport::Channel;
 use tonic::Request;
+
+use proto_buf::combiner::linear_combiner_client::LinearCombinerClient;
+use proto_buf::combiner::{LtBatch, LtHistoryBatch};
+use proto_buf::transformer::transformer_client::TransformerClient;
+use proto_buf::transformer::{EventBatch, TermBatch};
 
 const BATCH_SIZE: u32 = 1000;
 const INTERVAL_SECS: u64 = 5;
