+++ linear-combiner/src/main.rs
@@ -1,21 +1,21 @@
-use error::LcError;
-use managers::{
-	checkpoint::CheckpointManager, index::IndexManager, item::ItemManager, mapping::MappingManager,
-	update::UpdateManager,
-};
-use proto_buf::{
-	combiner::{
-		linear_combiner_server::{LinearCombiner, LinearCombinerServer},
-		LtBatch, LtHistoryBatch, LtObject, Mapping, MappingQuery,
-	},
-	common::Void,
-	transformer::TermObject,
-};
-use rocksdb::{Options, DB};
 use std::error::Error;
+
+use rocksdb::{Options, DB};
 use tokio::sync::mpsc::channel;
 use tokio_stream::wrappers::ReceiverStream;
 use tonic::{transport::Server, Request, Response, Status, Streaming};
+
+use proto_buf::combiner::linear_combiner_server::{LinearCombiner, LinearCombinerServer};
+use proto_buf::combiner::{LtBatch, LtHistoryBatch, LtObject, Mapping, MappingQuery};
+use proto_buf::common::Void;
+use proto_buf::transformer::TermObject;
+
+use crate::error::LcError;
+use crate::managers::checkpoint::CheckpointManager;
+use crate::managers::index::IndexManager;
+use crate::managers::item::ItemManager;
+use crate::managers::mapping::MappingManager;
+use crate::managers::update::UpdateManager;
 
 pub mod error;
 pub mod item;
