+++ linear-combiner/src/managers/index.rs
@@ -1,5 +1,6 @@
-use crate::error::LcError;
 use rocksdb::DB;
+
+use crate::error::LcError;
 
 #[derive(Debug)]
 pub struct IndexManager;
@@ -26,8 +27,9 @@ impl IndexManager {
 
 #[cfg(test)]
 mod test {
-	use crate::managers::index::IndexManager;
 	use rocksdb::{Options, DB};
+
+	use super::*;
 
 	#[test]
 	fn should_update_and_get_index() {
