+++ linear-combiner/src/managers/item.rs
@@ -1,6 +1,7 @@
 use rocksdb::DB;
 
-use crate::{error::LcError, item::LtItem};
+use crate::error::LcError;
+use crate::item::LtItem;
 
 #[derive(Debug)]
 pub struct ItemManager;
@@ -56,8 +57,11 @@ impl ItemManager {
 
 #[cfg(test)]
 mod test {
-	use crate::{item::LtItem, managers::item::ItemManager};
 	use rocksdb::{Options, DB};
+
+	use crate::item::LtItem;
+
+	use super::*;
 
 	#[test]
 	fn should_update_item() {
